% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_barchart_module.R
\name{admin_barchart_module_ui}
\alias{admin_barchart_module_ui}
\title{Admin Barchart Module UI}
\usage{
admin_barchart_module_ui(id)
}
\arguments{
\item{id}{A shiny id}
}
\description{
Admin Barchart Module UI
}
