% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_datatable_module.R
\name{admin_datatable_module_server}
\alias{admin_datatable_module_server}
\title{Admin Data Table Module Server}
\usage{
admin_datatable_module_server(id, data, input_config)
}
\arguments{
\item{id}{A shiny id}

\item{data}{A shiny::reactive that returns a data frame}

\item{input_config}{A shiny::reactive that returns a named list or Null.}
}
\description{
Admin Data Table Module Server
}
