% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_datatable_module.R
\name{admin_datatable_module_ui}
\alias{admin_datatable_module_ui}
\title{Admin Data Table Module UI}
\usage{
admin_datatable_module_ui(id)
}
\arguments{
\item{id}{A shiny id}
}
\description{
Admin Data Table Module UI
}
