% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_module.R
\name{admin_module_server}
\alias{admin_module_server}
\title{Admin Module Server}
\usage{
admin_module_server(id, data)
}
\arguments{
\item{id}{A shiny id}

\item{data}{a shiny::reactive that returns a named list of data frames}
}
\description{
Admin Module Server
}
