% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_piechart_module.R
\name{admin_piechart_module_ui}
\alias{admin_piechart_module_ui}
\title{Admin Piechart Module UI}
\usage{
admin_piechart_module_ui(id)
}
\arguments{
\item{id}{A shiny id}
}
\description{
Admin Piechart Module UI
}
