% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_selection_module.R
\name{attribute_selection_module_server}
\alias{attribute_selection_module_server}
\title{Attribute Selection Module Server}
\usage{
attribute_selection_module_server(
  id,
  config,
  attribute_name,
  attribute_choices,
  attribute_input_default = shiny::reactive(NA),
  ui_label = shiny::reactive("Select attribute")
)
}
\arguments{
\item{id}{A shiny id}

\item{config}{A shiny::reactive that returns a named list or Null.}

\item{attribute_name}{A shiny::reactive that returns a string}

\item{attribute_choices}{A shiny::reactive that returns a named list}

\item{attribute_input_default}{A shiny::reactive that returns a string or NA}

\item{ui_label}{A shiny::reactive that returns a string}
}
\description{
Attribute Selection Module Server
}
