% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_selection_module.R
\name{attribute_selection_module_ui}
\alias{attribute_selection_module_ui}
\title{Attribute Selection Module UI}
\usage{
attribute_selection_module_ui(id)
}
\arguments{
\item{id}{A shiny id}
}
\description{
Attribute Selection Module UI
}
