% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_module.R
\name{barchart_module_server}
\alias{barchart_module_server}
\title{Barchart Module Server}
\usage{
barchart_module_server(id, data, config)
}
\arguments{
\item{id}{A shiny id}

\item{data}{A data frame}

\item{config}{A named list. All values must be columns in data:
- "name": shinydashboard::box box title
- "x_attribute": Column in data param (required)
- "color_attribute": Column in data param
- "group_attribute": Column in data param}
}
\description{
Barchart Module Server
}
