% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_functions.R
\name{create_barchart_config}
\alias{create_barchart_config}
\title{Create Barchart Config}
\usage{
create_barchart_config(entity, x_attribute, group_attribute, color_attribute)
}
\arguments{
\item{entity}{A string}

\item{x_attribute}{A stirng}

\item{group_attribute}{A string}

\item{color_attribute}{A string}
}
\description{
Create Barchart Config
}
