% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{create_extraneous_attribute_string}
\alias{create_extraneous_attribute_string}
\title{Create Extraneous Attribute String}
\usage{
create_extraneous_attribute_string(
  config_names,
  allowed_attributes,
  error_prefix = "Config malformed: "
)
}
\arguments{
\item{config_names}{A list of names from the config}

\item{allowed_attributes}{A list of attributes allowed for the config}

\item{error_prefix}{The prefix for the error message}
}
\description{
Create Extraneous Attribute String
}
