% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{create_missing_attr_msg}
\alias{create_missing_attr_msg}
\title{Create Missing Attribute Message}
\usage{
create_missing_attr_msg(
  config_names,
  required_attrbutes,
  error_prefix = "Config malformed: "
)
}
\arguments{
\item{config_names}{A list of names in the config}

\item{required_attrbutes}{The attributes the config must have}

\item{error_prefix}{The prefix for the error message}
}
\description{
Create Missing Attribute Message
}
