% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechart_functions.R
\name{create_piechart}
\alias{create_piechart}
\title{Create Piechart}
\usage{
create_piechart(data, config, palette = "viridis")
}
\arguments{
\item{data}{A data frame}

\item{config}{A named list. All values must be columns in data:
- "label_attribute" (required)}

\item{palette}{A name of a viridis() palette}
}
\description{
Create Piechart
}
