% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechart_functions.R
\name{create_piechart_config}
\alias{create_piechart_config}
\title{Create Piechart Config}
\usage{
create_piechart_config(label_attribute)
}
\arguments{
\item{label_attribute}{A stirng}
}
\description{
Create Piechart Config
}
