% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_plot_config}
\alias{create_plot_config}
\title{Create Plot Config}
\usage{
create_plot_config(attributes, plot_type)
}
\arguments{
\item{attributes}{A named list}

\item{plot_type}{A string}
}
\description{
Create Plot Config
}
