% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_functions.R
\name{create_plotly_barchart}
\alias{create_plotly_barchart}
\title{Create Plotly Barchart}
\usage{
create_plotly_barchart(
  plot_data,
  x_col = "x",
  y_col = "y",
  color_col = x_col,
  key_col = x_col,
  text_col = x_col,
  xlab = "",
  ylab = "",
  title = "",
  bar_colors = NULL,
  showlegend = TRUE
)
}
\arguments{
\item{plot_data}{A dataframe}

\item{x_col}{A string}

\item{y_col}{A string}

\item{color_col}{A string or NA}

\item{key_col}{A string or NA}

\item{text_col}{A string or NA}

\item{xlab}{A string}

\item{ylab}{A string}

\item{title}{A string}

\item{bar_colors}{A string or NULL}

\item{showlegend}{True or False}
}
\description{
Create Plotly Barchart
}
