% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechart_functions.R
\name{create_plotly_piechart}
\alias{create_plotly_piechart}
\title{Create Plotly Piechart}
\usage{
create_plotly_piechart(
  plot_data,
  labels_col = "labels",
  values_col = "values",
  palette = "viridis"
)
}
\arguments{
\item{plot_data}{A dataframe}

\item{palette}{A name of a viridis() palette}

\item{x_col}{A string}

\item{y_col}{A string}

\item{color_col}{A string or NA}

\item{key_col}{A string or NA}

\item{text_col}{A string or NA}

\item{xlab}{A string}

\item{ylab}{A string}

\item{title}{A string}

\item{showlegend}{True or False}
}
\description{
Create Plotly Piechart
}
