% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_functions.R
\name{create_stacked_barchart}
\alias{create_stacked_barchart}
\title{Create StackedBarchart}
\usage{
create_stacked_barchart(data, config)
}
\arguments{
\item{data}{A dataframe}

\item{config}{A named list}
}
\description{
Create StackedBarchart
}
