% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_module.R
\name{display_module_server}
\alias{display_module_server}
\title{Display Module Server}
\usage{
display_module_server(id, config, data)
}
\arguments{
\item{id}{A shiny id}

\item{config}{A shiny::reactive that returns a named list:
- "type": one of c("datatable", "barchart")
- "entity": a name in data
-  One of ("barchart", "datatable"). See barchart_module_server() and
datatable_module_server()}

\item{data}{A shiny::reactive that returns a named list of data frames.}
}
\description{
Display Module Server
}
