% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_plot_table}
\alias{get_plot_table}
\title{Get Plot Table}
\usage{
get_plot_table(type = NULL)
}
\arguments{
\item{type}{A string}
}
\description{
This table has various data used specifically for each plot type
}
