% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_plot_type}
\alias{get_plot_type}
\title{Get Plot Type}
\usage{
get_plot_type(config)
}
\arguments{
\item{config}{A named list with attribute name of "plot_type"}
}
\description{
Determines the plot type from the config
}
