% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_value_from_list}
\alias{get_value_from_list}
\title{Get Value From List}
\usage{
get_value_from_list(lst, value, default = NA)
}
\arguments{
\item{lst}{A named list}

\item{value}{The name of the value}

\item{default}{The default return value}
}
\description{
Gets the value from the list, or returns the default value if unable
}
