% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_viridis_colors_from_tbl}
\alias{get_viridis_colors_from_tbl}
\title{Get Viridis Colors From Tbl}
\usage{
get_viridis_colors_from_tbl(table, pallete = "viridis", column = "color")
}
\arguments{
\item{table}{A dataframe with column "color"}

\item{pallete}{A name of a viridis() pallete}

\item{column}{A name fo a column}
}
\description{
Get Viridis Colors From Tbl
}
