% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_module.R
\name{plotly_module_server}
\alias{plotly_module_server}
\title{Plotly Module Server}
\usage{
plotly_module_server(
  id,
  config,
  data,
  plot_function,
  required_config_attrbutes = shiny::reactive(c()),
  optional_config_attributes = shiny::reactive(c()),
  do_plot = shiny::reactive(TRUE)
)
}
\arguments{
\item{id}{A shiny id}

\item{config}{A shiny::reactive that returns a named list:
- "x_attribute": Column in data param (required)
- "color_attribute": Column in data param
- "group_attribute": Column in data param}

\item{data}{A shiny::reactive that returns A data frame.}

\item{do_plot}{A shiny::reactive that returns a logical.}
}
\description{
Plotly Module Server
}
