% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{validate_allowed_attributes}
\alias{validate_allowed_attributes}
\title{Validate Allowed Attributes}
\usage{
validate_allowed_attributes(
  config,
  required_attrbutes,
  optional_attributes,
  error_prefix = "Config malformed: "
)
}
\arguments{
\item{config}{The config being tested, a named list}

\item{required_attrbutes}{The attributes the config must have}

\item{optional_attributes}{The attributes the config is allowed to have}

\item{error_prefix}{The prefix for the error message}
}
\description{
Validate Allowed Attributes
}
