% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_selection_functions.R
\name{validate_attribute_choices}
\alias{validate_attribute_choices}
\title{Validate Attribute Choices}
\usage{
validate_attribute_choices(choices)
}
\arguments{
\item{choices}{A}
}
\description{
Checks that the choices have names and are at least length 1
}
