% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_selection_functions.R
\name{validate_attribute_config}
\alias{validate_attribute_config}
\title{Validate Attribute Config}
\usage{
validate_attribute_config(config)
}
\arguments{
\item{config}{A named list}
}
\description{
Checks that the config has names and has at least one value.
}
