% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_selection_functions.R
\name{validate_attribute_name}
\alias{validate_attribute_name}
\title{Validate Attribute Name}
\usage{
validate_attribute_name(name)
}
\arguments{
\item{name}{A string}
}
\description{
Checks that the name is a length one character vector
}
