% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_module_functions.R
\name{validate_plot_config}
\alias{validate_plot_config}
\title{Validate Plot Config}
\usage{
validate_plot_config(config, error_prefix = "Config malformed: ")
}
\arguments{
\item{config}{The config being tested, a named list}

\item{error_prefix}{The prefix for the error message}
}
\description{
Validate Plot Config
}
