% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{validate_plotly_config}
\alias{validate_plotly_config}
\title{Validate Plotly Config}
\usage{
validate_plotly_config(
  config,
  required_attrbutes = c(),
  optional_attributes = c(),
  error_prefix = "Config malformed: "
)
}
\arguments{
\item{config}{The config being tested, a named list}

\item{required_attrbutes}{The attributes the config must have}

\item{optional_attributes}{The attributes the config is allowed to have}

\item{error_prefix}{The prefix for the error message}
}
\description{
Validate Plotly Config
}
