% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{validate_plotly_data}
\alias{validate_plotly_data}
\title{Validate Plotly Data}
\usage{
validate_plotly_data(data, error_prefix = "Data malformed: ")
}
\arguments{
\item{data}{A dataframe}

\item{error_prefix}{The prefix for the error message}
}
\description{
Validate Plotly Data
}
