% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{validate_required_attributes}
\alias{validate_required_attributes}
\title{Validate Required Attributes}
\usage{
validate_required_attributes(
  config,
  required_attrbutes,
  error_prefix = "Config malformed: "
)
}
\arguments{
\item{config}{The config being tested, a named list}

\item{required_attrbutes}{The attributes the config must have}

\item{error_prefix}{The prefix for the error message}
}
\description{
Validate Required Attributes
}
