% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_barchart_module.R
\name{admin_barchart_module_server}
\alias{admin_barchart_module_server}
\title{Admin Barchart Module Server}
\usage{
admin_barchart_module_server(id, data, input_config, entity)
}
\arguments{
\item{id}{A shiny id}

\item{data}{A shiny::reactive that returns a data frame}

\item{input_config}{A shiny::reactive that returns a named list or Null.}

\item{entity}{A shiny::reactive that returns a string}
}
\description{
Admin Barchart Module Server
}
