% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_barchart_module.R
\name{admin_barchart_module_server}
\alias{admin_barchart_module_server}
\title{Admin Barchart Module UI}
\usage{
admin_barchart_module_server(id, data, entity)
}
\arguments{
\item{id}{A shiny id}

\item{data}{A data frame}

\item{entity}{A string}
}
\description{
Admin Barchart Module UI
}
