% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin_datatable_module.R
\name{admin_datatable_module_server}
\alias{admin_datatable_module_server}
\title{Admin Data Table Module UI}
\usage{
admin_datatable_module_server(id, data, entity)
}
\arguments{
\item{id}{A shiny id}

\item{data}{A data frame}

\item{entity}{A string}
}
\description{
Admin Data Table Module UI
}
