% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_module.R
\name{barchart_module_server}
\alias{barchart_module_server}
\title{Barchart Module Server}
\usage{
barchart_module_server(id, config, data, do_plot = shiny::reactive(T))
}
\arguments{
\item{id}{A shiny id}

\item{config}{A shiny::reactive that returns a named list:
- "x_attribute": Column in data param (required)
- "color_attribute": Column in data param
- "group_attribute": Column in data param}

\item{data}{A shiny::reactive that returns A data frame.}

\item{do_plot}{A shiny::reactive that returns a logical.}
}
\description{
Barchart Module Server
}
