% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_functions.R
\name{create_grouped_barchart}
\alias{create_grouped_barchart}
\title{Create Grouped Barchart}
\usage{
create_grouped_barchart(data, config)
}
\arguments{
\item{data}{A dataframe}

\item{config}{A named list}
}
\description{
Create Grouped Barchart
}
