% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piechart_functions.R
\name{create_plotly_piechart}
\alias{create_plotly_piechart}
\title{Create Plotly Piechart}
\usage{
create_plotly_piechart(
  plot_data,
  labels_col = "labels",
  values_col = "values",
  palette = "viridis"
)
}
\arguments{
\item{plot_data}{A dataframe}

\item{labels_col}{A name of the column}

\item{values_col}{A name of a column}

\item{palette}{A name of a viridis() palette}
}
\description{
Create Plotly Piechart
}
