% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_functions.R
\name{create_stacked_grp_barchart}
\alias{create_stacked_grp_barchart}
\title{Create Stacked Grouped Barchart}
\usage{
create_stacked_grp_barchart(data, config, palette = "viridis")
}
\arguments{
\item{data}{A dataframe}

\item{config}{A named list. All values must be columns in data:
- "x_attribute"
- "color_attribute"
- "group_attribute"}

\item{palette}{A name of a viridis() palette}
}
\description{
Create Stacked Grouped Barchart
}
