% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barchart_functions.R
\name{create_standard_barchart}
\alias{create_standard_barchart}
\title{Create Standard Barchart}
\usage{
create_standard_barchart(data, config)
}
\arguments{
\item{data}{A data frame}

\item{config}{A named list. All values must be columns in data:
- "x_attribute" (required)}
}
\description{
Create Standard Barchart
}
