% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_module.R
\name{datatable_module_server}
\alias{datatable_module_server}
\title{Data Table Module UI}
\usage{
datatable_module_server(id, config, data)
}
\arguments{
\item{id}{A shiny id}

\item{config}{A shiny::reactive that returns a named list:
- "name": a title for the shinydashboard::box}

\item{data}{A shiny::reactive that returns A data frame.}
}
\description{
Data Table Module UI
}
