% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_module.R
\name{datatable_module_server}
\alias{datatable_module_server}
\title{Data Table Module UI}
\usage{
datatable_module_server(id, config, data, do_plot = shiny::reactive(T))
}
\arguments{
\item{id}{A shiny id}

\item{config}{A shiny::reactive that returns a list}

\item{data}{A shiny::reactive that returns A data frame.}

\item{do_plot}{A shiny::reactive that returns a logical.}
}
\description{
Data Table Module UI
}
