% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_module.R
\name{datatable_module_server}
\alias{datatable_module_server}
\title{Data Table Module UI}
\usage{
datatable_module_server(id, data, ...)
}
\arguments{
\item{id}{A shiny id}

\item{data}{A shiny::reactive that returns A data frame.}

\item{...}{unused arguments}
}
\description{
Data Table Module UI
}
