% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable_module.R
\name{datatable_module_server}
\alias{datatable_module_server}
\title{Data Table Module UI}
\usage{
datatable_module_server(id, config, data)
}
\arguments{
\item{id}{A shiny id}

\item{config}{A named list. All values must be columns in data:
- "name": shinydashboard::box box title}

\item{data}{A data frame}
}
\description{
Data Table Module UI
}
