% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_module.R
\name{plotly_module_server}
\alias{plotly_module_server}
\title{Plotly Module Server}
\usage{
plotly_module_server(
  id,
  config,
  data,
  plot_function,
  required_config_attrbutes = shiny::reactive(c()),
  optional_config_attributes = shiny::reactive(c())
)
}
\arguments{
\item{id}{A shiny id}

\item{config}{A shiny::reactive that returns a named list}

\item{data}{A shiny::reactive that returns A data frame.}

\item{plot_function}{A function that plots the table}

\item{required_config_attrbutes}{Attributes thatare required in the config}

\item{optional_config_attributes}{Attributes that are allowed in the config}
}
\description{
Plotly Module Server
}
