% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_module.R
\name{plotly_module_server}
\alias{plotly_module_server}
\title{Plotly Module Server}
\usage{
plotly_module_server(id, data, config, plot_type, ...)
}
\arguments{
\item{id}{A shiny id}

\item{data}{A shiny::reactive that returns A data frame.}

\item{config}{A shiny::reactive that returns a named list}

\item{plot_type}{The type of plot, see get_plot_table()$plot_type}

\item{...}{unused arguments}
}
\description{
Plotly Module Server
}
